Fixes to active app, quietmode handling and radio on event.
1) Don't assume 3GPP as active app on CDMA with LTE device

IccCardProxy assumes, that on a CDMA+LTE device, the active app
is the SIM/USIM and not the RUIM/CSIM. However, for voice RUIM/CSIM
is the active app on such a device. All services, like FDN,etc. should
be tied to the active voice app.
CRs-Fixed: 428304, 506938

2) Uicc: Do not be in QuietMode if the device is LTE capable

The voice radio tech can be set either due to a VOICE_RADIO_TECH_CHANGED
unsol or due to a change in the radio technology in the
VOICE_REGISTRATION_STATE response. Ideally when the preferred network
mode on the UI changes, the radioTech update received from both the
messages should belong to the same radio tech family (3GPP or 3GPP2).
But if they do not, (for example, the system reports different radioTechs
in both the messages for a few seconds due to a delay in updating the
system state), an incorrect radioTech which may not be supported on
the current network mode will be set by IccCardProxy. Hence
incorrect value of the quietmode will be set leading to broadcasting
incorrect sim state changes. For example, in a LTE only network mode,
if a radioTech from 3GPP2 family is reported, QuietMode would be set to
true and hence sim state changes will not be broadcasted.

If the device is LTE capable, always be out of QuietMode
and therefore broadcast any sim state changes. For this,
introduce a new config item which indicates if the device is LTE
capable or not (true value indicating that the device is LTE capable).
CRs-Fixed: 628506

3) Uicc: Update icc card state when radio is turned on.

When the radio is off, if the icc card status changes,
the icc card state will be updated. At a later point when
the radio turns on, if the icc card status change event is not
received, ,the card state will not be restored and an incorrect
card state may be present in IccCardProxy.

Always update the icc card state (externalState) when the
radio turns on.
CRS-Fixed: 646808

Change-Id: I70894a1aab0a812a50f1cbab986161bb3b708b85

Conflicts:
	src/java/com/android/internal/telephony/uicc/IccCardProxy.java