Stk: Add support for alpha in REFRESH command
RIL_UNSOL_STK_EVENT_NOTIFY is given by RIL in case there is alpha
id to be displayed to user in refresh proactive cmd.
Support is added to display Alpha id to user when this event comes
due to REFRESH proactive cmd.
Also if refresh mode is init or reset, idle mode text needs to be
cleared as per stk 3GPP spec 11.14. This logic is handled in stk
app to satisfy both the above requiremnts.

CRs-Fixed: 474693

Change-Id: Ic16da08d611753f2ee4a5b288f46b985078ca46f

Conflicts:
	src/java/com/android/internal/telephony/cat/CatCmdMessage.java