Fix back to back data call requests on same apn.
If one ApnContext is in DISCONNECTING state and another ApnContext
which uses the same apn needs to be brought up, current code doesnot
share the same DataConnection object. This may lead to back to back
requests on same apn. To avoid this, allow sharing of DC even when it
is in DISCONNECTING state. This will allow DC to be brought up immediately
after getting disconnected.

CRs-Fixed: 712333
Change-Id: I3d6fef87014f339b04393ec33f86d89beab8c644