DSDA: Update service state to POWER_OFF in airplane mode
When emergency call attempted, in telephony service it is
service state is checked, if it is POWER_OFF, then modem is brought
from LPM to online mode and call is attempted. However if service
state is not POWER_OFF, dial request is attempted.

In current implementation, due to IWLAN feature enablement, polling
is also done in airplane mode, which is leading to setting of service state
"out of service" instead of POWER_OFF. As service state is not POWER_OFF,
when ecall is attempted, telephony service doesn't bring modem out of LPM and
directly places dial request which fails expectedly. And hence error dialog shown
to user instead of succesfully placing the ecall.

Current change takes care that, in airplane mode, if polling happens,
set service stateto POWER_OFF regardless of response received for
ril request VOICE_REGISTRATION_STATE.

Change-Id: I000c1dda446a684c61d2b806caa345f2771d7a08
CRs-Fixed: 722203

Conflicts:
	src/java/com/android/internal/telephony/cdma/CdmaServiceStateTracker.java