SmsUsageMonitor: update SMS limits with ContentObserver
Previously if a user exceeded the SMS limit and then chose to
increase the limit, the alert dialog continues to show every
new message the user attempts to send. This patch will update
the SMS limit variables when the limits are updated so that
when the user increases the SMS limit, the user will not
continue to see the alert dialog.

Change-Id: I6e10c5c5e8633f12ad2280c07d55cf423e740267